+++ web/src/App.js
@@ -6,8 +6,6 @@ import Footer from "./components/base/Footer";
 import Landing from "./components/landingpage/Landing";
 import Travelinnovation from "./components/ninedots/travelinnovation";
 import Assortment from './components/assortment/Assortment';
-import Home from "./components/home/Accommodations";
-import Booking from "./components/booking/Booking";
 import About from "./components/about/About";
 import Whydomits from "./components/about/Whydomits";
 import Release from "./components/about/release.js";
@@ -23,25 +21,16 @@ import HostListings from "./components/hostdashboard/HostListings";
 import HostCalendar from "./components/hostdashboard/HostCalendar";
 import HostSettings from "./components/hostdashboard/HostSettings";
 import ListingDetails from './components/listingdetails/ListingDetails';
-import DynamicListing from './components/listingdetails/dynamicListingDetails.js';
-import HomeDashboard from "./components/admindashboard/HomeDashboard";
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import React, { useEffect, useState } from "react";
 import Modal from "react-modal";
-import BookingOverview from "./components/bookingprocess/BookingOverview";
-import BookingPayment from "./components/bookingprocess/BookingPayment";
-import BookingConfirmed from "./components/bookingprocess/BookingConfirmed";
-import BookingDeclined from "./components/bookingprocess/BookingDeclined";
 import GuestDashboard from './components/guestdashboard/GuestDashboard';
-import { ProtectedRoute } from "./components/protectedroute/ProtectedRoute.tsx";
 import Disclaimers from "./components/disclaimers/Disclaimers";
 import Policy from "./components/disclaimers/Policy";
 import Terms from "./components/disclaimers/Terms";
 import Login from "./components/base/Login";
 import Register from "./components/base/Register";
 import ConfirmRegister from "./components/base/ConfirmRegister";
-import Error from "./components/errorpage/errorpage";
-import Stripe from 'stripe';
 import { AuthProvider } from './components/base/AuthContext';
 import PaymentsGuestDashboard from "./components/guestdashboard/PaymentsGuestDashboard";
 import Chat from "./components/chat/Chat";
@@ -79,7 +68,6 @@ function App() {
                         {currentPath !== '/admin' && <Header setSearchResults={setSearchResults} />}
                         <Routes>
                             <Route path="/" element={<Assortment searchResults={searchResults} />} />
-                            <Route path="/home" element={<Home />} />
                             <Route path="/about" element={<About />} />
                             <Route path="/release" element={<Release />} />
                             <Route path="/helpdesk" element={<Helpdesk />} />
@@ -91,14 +79,7 @@ function App() {
                             <Route path="/login" element={<Login />} />
                             <Route path="/register" element={<Register />} />
                             <Route path="/confirm-email" element={<ConfirmRegister />} />
-                            <Route path="/error" element={<Error />} />
-                            <Route path="/booking" element={<Booking />} />
                             <Route path="/listingdetails" element={<ListingDetails />} />
-                            <Route path="/dynamicListing" element={<DynamicListing />} />
-                            <Route path="/bookingoverview" element={<BookingOverview />} />
-                            <Route path="/bookingpayment" element={<BookingPayment />} />
-                            <Route path="/bookingconfirmed" element={<BookingConfirmed />} />
-                            <Route path="/bookingdeclined" element={<BookingDeclined />} />
 
                             {/* Chat */}
                             <Route path="/chat" element={<Chat />} />
@@ -127,7 +108,6 @@ function App() {
                             <Route path="/policy" element={<Policy />} />
                             <Route path="/terms" element={<Terms />} />
                             <Route path="/disclaimers" element={<Disclaimers />} />
-                            <Route path="/admin" element={<ProtectedRoute allowedRoles={["Admin"]} page={<HomeDashboard />} redirectTo='/' />} />
                         </Routes>
                         {renderFooter()}
                     </div>
