--- web/src/components/admindashboard/AdminDashboard.css
@@ -1,90 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400;500;600;700&display=swap');
-
-.manage-list {
-  position: relative;
-  left: 15px;
-  top: 10px;
-  padding-bottom: 20px;
-  color: #000;
-  font-family: "Quicksand";
-  font-size: 9px;
-  font-weight: 400;
-  line-height: normal;
-  letter-spacing: 1.5px;
-  text-transform: uppercase;
-  display: flex;
-  width: 90%;
-  flex-direction: column;
-  flex-shrink: 0;
-}
-
- .item-preview {
-  padding: 8px;
-  margin: 4px 0;
-  border-radius: 15px;
-   background-color: #FFFFFF;
-  box-shadow: 0px 2px 5px 0px rgba(0, 0, 0, 0.25);
-}
-
-.item-preview:hover{
-  transition: transform 0.2s ease;
-  transform: scale(0.95);
-}
-
-.list-button {
-  line-height: normal;
-  letter-spacing: 1.5px;
-  text-transform: uppercase;
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  background-color: #ffffff;
-}
-.navbar {
-  padding: 25px;
-  display: flex;
-  align-items: center;
-}
-.navbar .img{
-  position: relative;
-}
-.navbar .nav {
-  margin-left: auto;
-  font-family: "Quicksand";
-  font-size: 16px;
-  font-weight: 600;
-  line-height: 0,5;
-  letter-spacing: 1.1px;
-  text-transform: uppercase;
-  text-align: center;
-  width: 265px;
-  height: 60px;
-  flex-shrink: 0;
-  border-radius: 15px;
-  padding-top: 5px;
-  box-shadow: 0px 2px 5px 0px rgba(0, 0, 0, 0.25);
-}
-
-.grid-container {
-  display: grid;
-  grid-template-columns: 20% 80%;
-  grid-template-areas:
-    'header header header header header header header header header header header header header header header header header header'
-    'menu main main main main main main main main main main main main main main main main main';
-}
-
-.header { grid-area: header; }
-.menu {
-  grid-area: menu;
-  border-radius: 15px;
-  background: #F4F4F4;
-  margin-left: 25px;
-}
-.main {
-  padding: 25px;
-  grid-area: main;
-  border-radius: 15px;
-  background: #F4F4F4;
-  margin-left: 15px;
-  margin-right: 25px;
-}
