--- web/src/components/booking/booking.css
@@ -1,1 +1,0 @@
-/*pagina is nog leeg*/
