--- web/src/components/bookingprocess/BookingConfirmedDeclined.css
@@ -1,150 +1,0 @@
-.view-booking-button button,
-.try-again-button button {
-    width: 100%;
-    height: 100%;
-    padding: 10px;
-    background: #0D9813;
-    border-radius: 15px;
-    justify-content: center;
-    align-items: center;
-    color: white;
-    font-size: 20px;
-    font-family: Kanit;
-    font-weight: 400;
-    line-height: 24px;
-    word-wrap: break-word;
-    margin-top: 40px;
-}
-
-.booking-confirmed-section {
-    padding: 20px;
-    margin-top: 20px;
-    width: 100%;
-    background: #0D9813;
-    color: white;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.10);
-    border-radius: 15px
-}
-
-.booking-confirmed-title {
-    color: white;
-    font-size: 18px;
-    font-family: Kanit;
-    font-weight: 400;
-    line-height: 24px;
-    word-wrap: break-word
-}
-
-.booking-confirmed-subtitle {
-    color: white;
-    font-size: 16px;
-    font-family: Kanit;
-    font-weight: 300;
-    line-height: 24px;
-    word-wrap: break-word
-}
-
-.booking-confirmed-section {
-    display: flex;
-    align-items: center; /* Vertically center content */
-}
-
-.booking-confirmed-content {
-    flex-grow: 1; /* Allow content to grow and fill remaining space */
-}
-
-.booking-confirmed-check {
-    margin-left: 10px; /* Add some space between content and checkmark */
-}
-
-.booking-supertitle {
-    color: black;
-    font-size: 24px;
-    font-family: Kanit;
-    font-weight: 400;
-    line-height: 24px;
-    word-wrap: break-word
-}
-
-.booking-confirmation-details {
-    margin-top: 20px;
-    margin-bottom: 20px;
-    color: black;
-    font-size: 16px;
-    font-family: Kanit;
-    font-weight: 200;
-    line-height: 24px;
-    word-wrap: break-word;
-    display: flex; /* Use flexbox */
-    justify-content: space-between; /* Align items with space between them */
-}
-
-.booking-confirmation-details a {
-    color: #0A760F;
-    font-size: 16px;
-    font-family: Kanit;
-    font-weight: 400;
-    text-decoration: underline;
-    line-height: 24px;
-    letter-spacing: 0.16px;
-    word-wrap: break-word;
-}
-
-.booking-confirmed-image-section {
-    margin-left: 5%;
-}
-
-.booking-declined-section {
-    padding: 20px;
-    margin-top: 20px;
-    margin-bottom: 20px;
-    background: white;
-    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.10);
-    border-radius: 15px;
-    border: 2px #FF0000 solid;
-}
-
-.booking-declined-title {
-    color: black;
-    font-size: 16px;
-    font-family: Kanit;
-    font-weight: 300;
-    line-height: 24px;
-    word-wrap: break-word;
-    margin-right: 10px; /* Add some space between title and subtitle */
-}
-
-.booking-declined-subtitle {
-    color: black;
-    font-size: 16px;
-    font-family: Kanit;
-    font-weight: 300;
-    line-height: 24px;
-    word-wrap: break-word;
-}
-
-.booking-declined-underline {
-    color: black;
-    font-size: 16px;
-    font-family: Kanit;
-    font-weight: 300;
-    line-height: 24px;
-    word-wrap: break-word;
-    margin-right: 0.5%;
-    text-decoration: underline;
-}
-
-
-.booking-declined-section {
-    display: flex;
-    align-items: center; /* Vertically center content */
-}
-
-.booking-declined-content {
-    flex-grow: 1; /* Allow content to grow and fill remaining space */
-}
-
-.booking-declined-alert {
-    margin-left: 10px; /* Add some space between content and checkmark */
-}
-
