--- web/src/components/errorpage/errorpage.css
@@ -1,29 +1,0 @@
-.errorpage {
-    font-size: x-large;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    height: 100vh; 
-}
-
-.errorpage h3{
-    width:600px;
-}
-
-.errorpage button{
-    background-color: #fff;
-    border: solid 1px #0D9813;
-    border-radius: 30px;
-    margin: 10px;
-    padding: 10px 20px;
-    backdrop-filter: blur(20px);
-    justify-content: center;
-    align-items: center;
-}
-
-.errorpage button:hover{
-    background: #cccccc;
-    cursor: pointer;
-}
