+++ web/src/components/guestdashboard/GuestDashboard.js
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react";
-import accommodationImg from "../../images/accommodationtestpic1.png";
 import editIcon from "../../images/icons/edit-05.png";
 import { useNavigate } from 'react-router-dom';
 import { API, graphqlOperation, Auth } from "aws-amplify";
@@ -68,7 +67,6 @@ const GuestDashboard = () => {
                         <div className="bookingBox">
                             <h3>Current Booking</h3>
                             <p>Tropical 12 person villa with pool</p>
-                            <img src={accommodationImg} alt="Booking" />
                             <p>Host: John Doe</p>
                         </div>
                         <div className="messageBoxes">
