+++ web/src/components/hostdashboard/Accommodations.js
@@ -3,37 +3,38 @@ import { Link, useNavigate } from "react-router-dom";
 import './Accommodations.css';
 
 const Accommodations = ({ searchResults }) => {
-    const [accolist, setAccolist] = useState([]);
-    const navigate = useNavigate();
+  const [accolist, setAccolist] = useState([]);
+  const navigate = useNavigate();
 
   const formatData = (items) => {
     return items.map((item) => ({
-      image: `https://accommodationphotos.s3.eu-north-1.amazonaws.com/${item.PhotoUrls}`,
+      image: item.Images.image1, // Use the image URLs from the Images object
       title: item.Title,
-      details: item.description, 
-      size: `${item.Size}m²`,
-      price: `€${item.Price} per night`,
-      id: item['#PK'], 
+      details: item.Description, // Use the Description property
+      size: `${item.Measurements}m²`, // Use the Measurements property
+      price: `€${item.Rent} per night`, // Use the Rent property
+      id: item.ID, // Use the ID property
       bathrooms: `${item.Bathrooms} Bathrooms`,
       bedrooms: `${item.Bedrooms} Bedrooms`,
-      persons: `${item.Persons} Persons`,
+      persons: `${item.Guests} Persons`, // Use the Guests property
     }));
   };
 
-  useEffect(() => {
+  useEffect(() => {  
     const fetchData = async () => {
       try {
-        const response = await fetch('https://cfeo8gr5y0.execute-api.eu-north-1.amazonaws.com/dev/accommodation');
+        const response = await fetch('https://6jjgpv2gci.execute-api.eu-north-1.amazonaws.com/dev/ReadAccommodation');
         if (!response.ok) {
           throw new Error('Failed to fetch data');
         }
         const responseData = await response.json();
-        setAccolist(formatData(responseData));
+        const data = JSON.parse(responseData.body); // Parse JSON from the body property
+        console.log(data);
+        setAccolist(formatData(data));
       } catch (error) {
         console.error('Error fetching or processing data:', error);
       }
     };
-
     if (searchResults && searchResults.length > 0) {
       setAccolist(formatData(searchResults));
     } else {
@@ -41,29 +42,29 @@ const Accommodations = ({ searchResults }) => {
     }
   }, [searchResults]);
 
+  const handleClick = (ID) => {
+    navigate(`/listingdetails?ID=${encodeURIComponent(ID)}`);
+  };
+
   return (
     <div id="card-visibility">
-      {accolist.length === 0 ? (
-        <div className="no-results">Geen accommodaties gevonden</div>
-      ) : (
-        accolist.map((accommodation, index) => (
-          <div className="accocard" key={index}>
-            <Link to={`/listingdetails/`} className="accocard-link">
-              <img src={accommodation.image} alt={accommodation.title} />
-              <div className="accocard-content">
-                <div className="accocard-title">{accommodation.title}</div>
-                <div className="accocard-price">{accommodation.price}</div>
-                <div className="accocard-detail">{accommodation.details}</div>
-                <div className="accocard-specs">
-                  <div className="accocard-size">{accommodation.size}</div>
-                  <div className="accocard-size">{accommodation.bathrooms}</div>
-                  <div className="accocard-size">{accommodation.bedrooms}</div>
-                </div>
+      {accolist.map((accommodation, index) => (
+        <div className="accocard" key={index} onClick={() => handleClick(accommodation.id)}>
+            <img src={accommodation.image} alt={accommodation.title} />
+            <div className="accocard-content">
+              <div className="accocard-title">{accommodation.title}</div>
+              <div className="accocard-price">{accommodation.price}</div>
+              <div className="accocard-detail">{accommodation.details}</div>
+
+              <div className="accocard-specs">
+                <div className="accocard-size">{accommodation.size}</div>
+                <div className="accocard-size">{accommodation.bathrooms}</div>
+                <div className="accocard-size">{accommodation.bedrooms}</div>
+                <div className="accocard-size">{accommodation.persons}</div>
               </div>
-            </Link>
-          </div>
-        ))
-      )}
+            </div>
+        </div>
+      ))}
     </div>
   );
 };
